+++ src/FreeElephants/JsonApiToolkit/Middleware/BodyParser.php
@@ -2,6 +2,8 @@
 
 namespace FreeElephants\JsonApiToolkit\Middleware;
 
+use Fig\Http\Message\StatusCodeInterface;
+use FreeElephants\JsonApiToolkit\Psr\JsonApiResponseFactory;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Server\MiddlewareInterface;
@@ -9,10 +11,28 @@ use Psr\Http\Server\RequestHandlerInterface;
 
 class BodyParser implements MiddlewareInterface
 {
+    private JsonApiResponseFactory $responseFactory;
+
+    public function __construct(JsonApiResponseFactory $responseFactory)
+    {
+        $this->responseFactory = $responseFactory;
+    }
+
     public function process(ServerRequestInterface $request, RequestHandlerInterface $handler): ResponseInterface
     {
         $request->getBody()->rewind();
-        $request = $request->withParsedBody(json_decode($request->getBody()->getContents(), true));
+
+        $decodedBody = json_decode($request->getBody()->getContents(), true);
+
+        if ($decodedBody === null && json_last_error() !== JSON_ERROR_NONE) {
+            return $this->responseFactory->createSingleErrorResponse(
+                'Provided json is invalid',
+                StatusCodeInterface::STATUS_BAD_REQUEST,
+                $request
+            );
+        }
+
+        $request = $request->withParsedBody($decodedBody);
 
         return $handler->handle($request);
     }