+++ tests/FreeElephants/JsonApiToolkit/Middleware/BodyParserTest.php
@@ -2,15 +2,18 @@
 
 namespace FreeElephants\JsonApiToolkit\Middleware;
 
+use Fig\Http\Message\StatusCodeInterface;
 use FreeElephants\JsonApiToolkit\AbstractHttpTestCase;
 use Psr\Http\Message\ServerRequestInterface;
 
 class BodyParserTest extends AbstractHttpTestCase
 {
+
     public function testProcess()
     {
         $request = $this->createServerRequest('POST', '/foo');
-        $request->getBody()->write(<<<JSON
+        $request->getBody()->write(
+            <<<JSON
 {
     "data": {
         "id": "foo",
@@ -19,18 +22,48 @@ class BodyParserTest extends AbstractHttpTestCase
 }
 JSON
         );
-        $handler = $this->createRequestHandlerWithAssertions(function (ServerRequestInterface $request) {
-            $this->assertSame([
-                'data' => [
-                    'id'   => 'foo',
-                    'type' => 'bar',
-                ],
-            ], $request->getParsedBody());
+        $handler = $this->createRequestHandlerWithAssertions(
+            function (ServerRequestInterface $request) {
+                $this->assertSame(
+                    [
+                        'data' => [
+                            'id'   => 'foo',
+                            'type' => 'bar',
+                        ],
+                    ],
+                    $request->getParsedBody()
+                );
 
+                return $this->createResponse();
+            }
+        );
+
+        $bodyParser = new BodyParser($this->createJsonApiResponseFactory());
+        $response = $bodyParser->process($request, $handler);
+
+        $this->assertResponseHasStatus($response, StatusCodeInterface::STATUS_OK);
+    }
+
+    public function testInvalidBody()
+    {
+        $request = $this->createServerRequest('POST', '/foo');
+        $request->getBody()->write(
+            <<<JSON
+{
+    "data": {
+        "id": "foo",
+        "type": "bar"
+}
+JSON
+        );
+        $handler = $this->createRequestHandlerWithAssertions(function () {
             return $this->createResponse();
         });
 
-        $bodyParser = new BodyParser();
-        $bodyParser->process($request, $handler);
+        $bodyParser = new BodyParser($this->createJsonApiResponseFactory());
+        $response = $bodyParser->process($request, $handler);
+
+        $this->assertResponseHasStatus($response, StatusCodeInterface::STATUS_BAD_REQUEST);
     }
+
 }